binder: Fix possible premature reaping of binder_proc
If a new node or new ref is created for a binder_proc
at the same time that the binder_proc becomes a zombie
it is possible that the binder_proc will be kfree'd
when there are still a node or ref with pointers to
the kfree'd proc. Avoid reaping the binder_proc if there
are any refs, nodes, or threads referring to them that
are still active.

Bug: 35301923
Change-Id: I062446c3499c78ca445f35fb8f409494125ddd9f
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>